Make module for TGWUI functions
